reorganised validation classes, changed phone validation
