+++ src/Controllers/FormActions/AddApplicantActionController.php
@@ -32,16 +32,8 @@ class AddApplicantActionController extends Controller
         $newApplicant = $request->getParsedBody();
 
         try {
-            ApplicantValidator::validate($newApplicant);
-            ApplicationValidator::validate($newApplicant);
-            ApplicationValidator::checkExists($newApplicant['circumstance_id'], $this->model->getAllCircumstances()[$newApplicant['circumstance_id'] - 1], "Circumstance ID");
-            ApplicationValidator::checkNumeric($newApplicant['circumstance_id'], "Circumstance ID");
-            ApplicationValidator::checkExists($newApplicant['funding_id'], $this->model->getAllFundingOptions()[$newApplicant['funding_id'] - 1], "Funding ID");
-            ApplicationValidator::checkNumeric($newApplicant['funding_id'], "Funding ID");
-            ApplicationValidator::checkExists($newApplicant['cohort_id'], $this->model->getAllCohorts()[$newApplicant['cohort_id'] - 1], "Cohort ID");
-            ApplicationValidator::checkNumeric($newApplicant['cohort_id'], "Cohort ID");
-            ApplicationValidator::checkExists($newApplicant['heard_about_id'], $this->model->getAllHearAboutUs()[$newApplicant['heard_about_id'] - 1], "Heard About ID");
-            ApplicationValidator::checkNumeric($newApplicant['heard_about_id'], "Heard About ID");
+            ApplicationValidator::validate($newApplicant, $this->model);
+
 
         } catch (Exception $e) {
             return $this->redirectWithError($response, '/admin/applicant/add', $e->getMessage());
